requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
psycopg2
python-decouple
Flask-bootstrap
Flask-Bootstrap
Flask-Login